breaks linkup.
start the bridge, then the switch some time afterwards. tuned to
reproduce failure on apb2.

This is using:
• BDB2-A
• ES2 bridges and switches

We were able to debug an issue with the link start up sequence failing. The
way to reproduce this is to start the bridge (device) then start the switch
(host). If the timing between the resets is within a rage (which is not
characterized) the link fails to start furthermore the host continues to
send TRG_UPR0 forever.

In summary there seems to be 2 issues identified:
• The bridge (device) does not restart its link startup timeout timer when
  the switch (host) starts the link startup sequence
• The switch (host) goes into a mode where it sends TRG_UPR0 forever in 1
  continues burst

Looking at trace LSS_FailureShowingLineReset.zip on Google Drive Eng
Workstreams -> Spiral 2 Workstreams -> SW – Firmware -> Link Debug
1. Starting at the marker “Host Start LSS” we can see the host send 2
   TRG_UPR0 which the device sees
2. The Device moves to Phase 1 (note only 2 TRG_UPR0’s should be sent not 3
   before moving to phase 1)
3. Going to Marker “Device End LSS” we can see the Device link start up time
   expires @ 92.04ms after starting (marker “Device Restart LSS”)
4. 188us after the device quits the host Moves from Phase 0 to Phase 2. The
   host continues to send TRG_UPR2
5. 7.26ms after the device quits (step 3) the device restarts LSS (marker
   “Device Start LSS #2”) at this marker we can also see the host quit sending
   TRG_UPR2
6. After this the host send TRG_UPR0 forever
7. The device continues to retry linkup

The other trace LSSFailure.zip is just another example of the failing
sequence.

We tested the reverse configuration where the bridge (device) was powered
after the Switch (host) there are particular timings where the device or
host timeout during the LSS but in all cases it seems to recover correctly.
There are a few traces in the HostFailsLSSandRecovers.zip. In this
configuration we did not see any problems we just are providing the traces
for your reference.