test app for toshiba
run:

Run the eng330 test component. The test program boots either the bridge
APB2 first, or the switch first, with a delay in between. It can
optionally sweep through a range of possible delays in 1ms increments,
or use a specified delay. Linkup status is checked and reported if
successful. Failed linkups will wait 20ms and retry 100 times.

Note, the integer values used here and reported from the program are
measured in microseconds, but do reflect wall clock time due to other
things going on in the program.

Usage:
    svc eng330 COMPONENT OPTION

COMPONENT is either 'bridge' or 'switch'. This selects which component
    to start first.

OPTION is either 'sweep,' or an integer between 1-500000.
    --If option is 'sweep,' it will sweep through the range in 1ms
    increments checking for linkup.
    --If option is an integer, it will use that time as the specified
    delay time and report linkup success/failure as above.

Examples:

1) Boot the bridge first, then the switch, sweeping through range
[0-~500ms]:
    svc eng330 bridge sweep

2) Boot switch first, then the bridge, sweeping through range
    svc eng330 switch sweep

3) Boot bridge first, then the switch, with a specified delay
    svc eng330 bridge 10000

4) Boot switch first, then the bridge, with a specified delay
    svc eng330 switch 138500