tsb_switch: es2: SRPT status report support
Implement FIFO status request transmission and response parsing. This is
used to retrieve the status of the SPI controller's FIFO in order to
determine useful things. Currently we only parse the available space in
the TX FIFO as well as the number of bytes available to read in the RX
FIFO.

Testing Done: send and receive data from svc to ap, check that computed
fifo sizes match expected values
Bug Number:
Review URL:
Reviewed by: