svc: wait for and handle ap module boot
Register the SVC for mailbox events and wait for the AP module to boot.
When the AP enumerates APBA, it will write into the switch-side mailbox,
notifying the SVC that it is ready, as well as it's position in the
endo.

HACK: Hotplug events are faked atm.

Testing Done:
Bug Number:
Review URL:
Reviewed by: